# fix compile error introduced by previous changes, we are not yet ready for this.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@1128 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/STEPFile.h
@@ -281,8 +281,8 @@ namespace STEP {
 		};
 
 		typedef PrimitiveDataType<int64_t>			INTEGER; 
-		typedef PrimitiveDataType<double>			REAL; 
-		typedef PrimitiveDataType<double>			NUMBER; 
+		typedef PrimitiveDataType<float>			REAL; 
+		typedef PrimitiveDataType<float>			NUMBER; 
 		typedef PrimitiveDataType<std::string>		STRING;
